Infrastructure - add repositories #43

+++ src/MonitorPet.Application/Model/PesoHistorico/ConsumptionModel.cs
@@ -4,5 +4,5 @@ public class ConsumptionModel
 {
     public DateTimeOffset Start { get; set; }
     public DateTimeOffset End { get; set; }
-    public decimal QttConsumption { get; set; }
+    public double QttConsumption { get; set; }
 }
